requirements.txt
pytesseract~=0.3.10
selenium~=4.8.0
Pillow~=9.4.0